# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Swagger OpenAPI documentation for Laravel API",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Fetch All Products",
                "description": "Retrieve a list of all products with details including reviews. Does not require authentication.",
                "operationId": "c5ad8f3c5bef0da1a3249908d4eccaf7",
                "responses": {
                    "200": {
                        "description": "Products Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "format": "uri",
                                                        "example": "https://picsum.photos/200/300"
                                                    },
                                                    "short_description": {
                                                        "type": "string",
                                                        "example": "desc"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 111.11
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.5,
                                                        "nullable": true
                                                    },
                                                    "discounted_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 99.999
                                                    },
                                                    "discount_percent": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10
                                                    },
                                                    "is_featured": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "reviews": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "rating": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to Retrieve Products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Single Product",
                "description": "Retrieve full details of a single product, including all images and reviews. Does not require authentication.",
                "operationId": "5204e328b97e4fdaaf21484e9c3508eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Product 1"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "product-1"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "desc"
                                                },
                                                "short_description": {
                                                    "type": "string",
                                                    "example": "desc"
                                                },
                                                "brand": {
                                                    "type": "string",
                                                    "example": "test"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 111.11
                                                },
                                                "discounted_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 99.999
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 99
                                                },
                                                "discount_percent": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "is_featured": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "primary_image": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "example": "https://picsum.photos/200/300"
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "format": "uri",
                                                        "example": "https://picsum.photos/200/300"
                                                    }
                                                },
                                                "reviews": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "rating": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "review": {
                                                                "type": "string",
                                                                "example": "desc"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "test"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-07-06T18:12:55.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to Retrieve Product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}